round2_requirements.txt
opencv-python==4.5.1.48
numpy==1.20.1
scikit-learn==0.24.1
matplotlib==3.3.4
Pillow==8.1.0
streamlit==0.77.0
